feat: unify style
