+++ app/market/[chainId]/[marketid]/RateChart.tsx
@@ -120,7 +120,7 @@ function RateChart({
   );
 
   return (
-    <Card className="bg-surface my-4 rounded-md p-4 shadow-sm">
+    <Card className="bg-surface my-4 rounded p-4 shadow-sm">
       <CardHeader className="flex items-center justify-between px-6 py-4 text-xl">
         <span />
         <ButtonGroup
